+++ app/portfolio/page.tsx
@@ -1,9 +1,8 @@
 "use client";
 
 import { useState } from "react";
-import { HistoricalCoin } from "@/interfaces/portfolio.interface";
+import { HistoricalCoin } from "@/interfaces/historicalCoin.interface";
 import PortfolioModal from "@/components/PortfolioModal";
-import PortfolioItem from "@/components/PortfolioItem";
 
 interface PortfolioItem {
   value: string;
@@ -15,8 +14,6 @@ const PortfolioPage = () => {
   const [showModal, setShowModal] = useState(false);
   const [historicalCoins, setHistoricalCoins] = useState<HistoricalCoin[]>([]);
   const [portfolioCoins, setPortfolioCoins] = useState<PortfolioItem[]>([]);
-
-  const hasCoins = historicalCoins.length > 0;
 
   return (
     <div className="dark:bg-[#09090c] bg-light-theme max-w-screen-2xl m-auto h-screen p-2 relative">
@@ -41,11 +38,6 @@ const PortfolioPage = () => {
       {showModal && (
         <PortfolioModal showModal={showModal} setShowModal={setShowModal} />
       )}
-      <PortfolioItem />
-      {/* {hasCoins &&
-        portfolioCoins.map((item, index) => (
-          <PortfolioItem key={index} item={item} />
-        ))} */}
     </div>
   );
 };
