+++ components/CoinGraphChart/index.tsx
@@ -68,6 +68,7 @@ const options = {
         display: true,
         color: "grey",
         maxTicksLimit: 10,
+        align: "inner",
       },
       border: {
         display: true,
