+++ components/PortfolioItem/index.tsx
@@ -1,10 +1,5 @@
-import { Portfolio } from "@/interfaces/portfolio.interface";
-
 const PortfolioItem = () => {
-  //   const coinData: Portfolio = portfolioCoins.find(
-  //     (item) => item.id === portfolio.coinId
-  //   )!;
-  return <div className="bg-red-500 w-[300px] h-[50px]">\</div>;
+  return <div className="bg-red-500 w-[300px] h-[50px]"></div>;
 };
 
 export default PortfolioItem;
