+++ interfaces/historicalCoin.interface.ts
@@ -1,0 +1,41 @@
+export interface HistoricalCoin {
+  id: string;
+  symbol: string;
+  name: string;
+  localization: { [key: string]: string };
+  image: {
+    thumb: string;
+    small: string;
+  };
+  market_data: {
+    current_price: { [currency: string]: number };
+    market_cap: { [currency: string]: number };
+    total_volume: { [currency: string]: number };
+  };
+  community_data: {
+    facebook_likes: number | null;
+    twitter_followers: number;
+    reddit_average_posts_48h: number;
+    reddit_average_comments_48h: number;
+    reddit_subscribers: number;
+    reddit_accounts_active_48h: string;
+  };
+  developer_data: {
+    forks: number;
+    stars: number;
+    subscribers: number;
+    total_issues: number;
+    closed_issues: number;
+    pull_requests_merged: number;
+    pull_request_contributors: number;
+    code_additions_deletions_4_weeks: {
+      additions: number;
+      deletions: number;
+    };
+    commit_count_4_weeks: number;
+  };
+  public_interest_stats: {
+    alexa_rank: number | null;
+    bing_matches: number | null;
+  };
+}
