+++ interfaces/searchCoin.interface.ts
@@ -1,0 +1,9 @@
+export interface SearchCoin {
+  id: string;
+  name: string;
+  api_symbol: string;
+  symbol: string;
+  market_cap_rank: number;
+  thumb: string;
+  large: string;
+}
