+++ utils/convertDate.tsx
@@ -1,0 +1,6 @@
+const convertDateFormat = (inputDate: string): string => {
+  const [year, month, day] = inputDate.split("-");
+  return `${day}-${month}-${year}`;
+};
+
+export default convertDateFormat;
